FaceRecognitionDjango/requirements.txt
certifi==2020.11.8
dlib==19.21.1
face-recognition==1.3.0	
face-recognition-models==0.3.0	
mysqlclient==2.0.1
numpy==1.19.4
opencv-python==4.4.0.46
pandas==1.1.4
Django==3.1.4
Pillow==8.0.1
PyMySQL==0.10.1
cv2-tools==2.4.0
